pyproject.toml
[project]
name = "hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.21",
    "pyyaml>=6.0.2",
]
